### 2023-07-27 #5690 @Kevin-Mattheus-Moerman
@cudmore I have just flagged this submission for a scope review by the editorial board. This is because this submission appears relatively small in terms of lines of code and I could use some help from the board to check if this work conforms to our [substantial scholarly effort criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This scope review usually takes about 2 weeks to complete.
