### 2023-09-06 #5690 @AoifeHughes
Hi @cudmore, I'm just putting out some requests to reviewers today, please feel free to make some suggestions of potential reviewers from your own knowledge or by using https://reviewers.joss.theoj.org to help us proceed faster
