### 2023-09-19 #5690 @AoifeHughes
@haesleinhuepf Would you be interested in helping to review this paper, JOSS review guidelines can [be found here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). I'd also be happy to directly answer any questions on our processes if you've not come across the journal before.
