### 2023-09-25 #5871 @editorialbot
[REVIEW]: Brightest path tracing: A Python package to trace the brightest path in 2D and 3D images
**Submitting author:** @cudmore (<a href="http://orcid.org/0000-0002-0440-1704">Robert Cudmore</a>)
**Repository:** https://github.com/mapmanager/brightest-path-lib
**Branch with paper.md** (empty if default branch): publish-joss-july-2023
**Version:** v1.0.15
