### 2023-10-01 #5871 @jingpengw
Generally speaking, this is a pretty valuable tool for the tracing of tubelike structures, such as neurons and blood vessels.
I would like to recommend to accept it.

There still exists some issues as follows.
# References
reference is not complete. There exist a large number of tracing algorithm and software tools. To my knowledge, most of them are for neuron and blood vessel tracing.
a few reviews are as follows:
Liu, Yufeng, et al. "Neuron tracing from light microscopy images: automation, deep learning and bench testing." Bioinformatics 38.24 (2022): 5329-5339.
Moccia, Sara, et al. "Blood vessel segmentation algorithms—review of methods, datasets and evaluation metrics." Computer methods and programs in biomedicine 158 (2018): 71-91.

Vaa3d is a widely used platform for neuron tracing. A variety of algorithms for neuron and blood vessel tracing are implemented as plugins.
Peng, Hanchuan, et al. "V3D enables real-time 3D visualization and quantitative analysis of large-scale biological image data sets." Nature biotechnology 28.4 (2010): 348-353.

# Performance
The authors claims that this tool is fast and efficient, but no benchmark is presented.
