### 2023-10-14 mapmanager/brightest-path-lib#43 @haesleinhuepf
Quick-start example lacks some import statement
Hi all,

I was just trying to execute the [quick start](https://mapmanager.net/brightest-path-lib/quickstart_guide/) example and retrieve an error in the second code cell.

code:
```
start = np.array([0,192])
goal = np.array([198,9])
```

Error:;
```
NameError                                 Traceback (most recent call last)
Cell In[3], line 1
----> 1 start = np.array([0,192])
      2 goal = np.array([198,9])

NameError: name 'np' is not defined
```

I recommend:
* Adding the `import numpy as np` statement to the example.
* Provide the example as notebook users can download and execute. We don't have to copy&paste it line by line then ;-)

Thanks!

Bestt,
Robert
