### 2023-10-14 mapmanager/brightest-path-lib#44 @haesleinhuepf
Quick start guide seems outdated
There is an example code line in the [quick start guide[(https://mapmanager.net/brightest-path-lib/quickstart_guide/) which seems outdated:

```
algorithm = brightest_path_lib.algorithm.AStarSearch(image, start, goal)
```

Error:
```
Cell In[6], line 1
----> 1 algorithm = brightest_path_lib.algorithm.AStarSearch(image, start, goal)

AttributeError: module 'brightest_path_lib' has no attribute 'algorithm'
```

Hint: I have this version installed:
```
import pkg_resources
pkg_resources.get_distribution("brightest-path-lib")
```
Output:
```
brightest-path-lib 1.0.12 (c:\users\haase\mambaforge\envs\nc1\lib\site-packages)
```

Installed like this:
```
mamba create --name nc1 python=3.9 napari=0.4.18 -c conda-forge
mamba activate nc1
pip install napari-tracing
```
