### 2023-10-14 #5871 @haesleinhuepf
Hi @cudmore and @VasudhaJha,

great paper! Unfortunately, I struggled using the library and the napari plugin, details below.

## Reproducibility

I tried to run the napari-plugin on my computer and retrieve an error message when attempting to start it. See:
* https://github.com/mapmanager/napari-tracing/issues/15

Also the [quick start guide]() seems not functional. I created separate github issues:
* https://github.com/mapmanager/brightest-path-lib/issues/43
* https://github.com/mapmanager/brightest-path-lib/issues/44

I could imagine that some issues are related to a conflict between published versions. On [pypi, brightest-path-lib](https://pypi.org/project/brightest-path-lib/) is published as version 1.0.12, on [github](https://github.com/mapmanager/brightest-path-lib) it says 1.0.15 is the current version.

## Documentation

In the [napari-plugin documentation](https://www.napari-hub.org/plugins/napari-tracing) I recommend adding screenshots in the user guide and also on top of the page as graphical abstract to attract users. Also consider providing an example data set to make sure first-time-users use it on a dataset where it works. Nothing is worse than trying some plugin for the first time and failing because you cannot find a suitable dataset. [Read more about user documentation](https://focalplane.biologists.com/2023/04/30/scientific-data-analysis-user-documentation-101/).

I'll continue the review once I can see it in action. I'm really looking forward to trying it out!

Thanks!

Best,
Robert
