### 2024-02-07 mapmanager/brightest-path-lib#44 @cudmore
Thanks for the comment, this has been fixed at least since version 1.0.16.

Sorry for the tardy response but we are back!!!
