### 2024-02-07 mapmanager/brightest-path-lib#43 @cudmore
Thanks for the feedback and tips. We have switched the quickstart to a Jupyter notebook so users can download and run and have added  `import numpy as np` to the notebook. All as you kindly suggested!
