### 2024-02-07 #5871 @cudmore
Hi @Kevin-Mattheus-Moerman. Yikes, I just noticed this last comment `If we do not hear back from you by January 26th we'll unfortunately have to reject this submission.`.

I have been making all the changes suggested by the reviewers and most (all) of our revisions got pushed to the repo in [v1.0.17 as of Oct 31, 2023](https://github.com/mapmanager/brightest-path-lib/releases/tag/v1.0.17).

I am still getting used to the JOSS review process which I am super optimistic about!. I did not respond in this review thread until now.

Can we still consider this manuscript open? I am now back on and responding to this thread.
