### 2024-05-25 #5871 @Kevin-Mattheus-Moerman
@cudmore I do hope you will be able to respond to the above soon. If we do not hear from you by Friday May 31st then we will unfortunately have to assume you are no longer interested in pursuing a publishing this work with JOSS.
