### 2024-05-31 #5871 @cudmore
Hi @Kevin-Mattheus-Moerman, @haesleinhuepf

My huge apologies for my silence. I am on the road (death in the family) and am responding today Fri 5/31 (EST).

As I mention, "The majority (all) of revisions for this submission are done". No excuses but a fairly typical situation where the student (Vasudha Jha) who implemented most of the code has left. We now have a replacement!

We have incorporated all reviewer suggested changes to the manuscript and are addressing technical issues that we propose **to be resolved and submit by Tue June 11th**. Please let me know if we can work with that timeframe.

1) We ran into a technical issue with distribution of the Windows version on PyPi. This was due to our attempt to speed up the code using Transonic (C++). We have chosen to roll back transonic and provide PyPi pip install without it. It works on macOS and we are checking if it works on Windows.

2) We ran into technical issue with keeping the napari plugin up to date because of breaking changes in Napari. We are currently running our code against the newest version of Napari and will have it running soon.

In general the content of the manuscript stands as it was originally submitted with the inclusion of all the reviewers suggestions. We just have to clear up purely technical (and important) issues with making sure the code runs on all platforms and that the Napari plugin is fully functional.

I kindly and humbly ask if we can "reset" this submission from Feb 2024. Let me know what you think for a **June 11th resubmission**.

Sincerely,

Robert
