### 2024-06-05 #5871 @Kevin-Mattheus-Moerman
@cudmore I feel it is best to withdraw this submission. JOSS does not publish static open source code, instead we aim to publish and promote open source projects that are actively been maintained and developed. In this case, given the delays so far, and the loss of a staff member (although a new one is being appointed), and the extent of the changes referred to, I believe it is best to withdraw this submission, and to resubmit on a future date.

We would welcome a re-submission once all of the above alluded to changes are made, AND once the project has sufficient active technical developers/contributors such that it is clear that the intention is to maintain it continuously, and such that requested changes can be worked on in a timely manner.
